Fixed bug - base64 TypeError: Incorrect padding
don't split the path with /,
use the full path and remove the last /

use urlsafe_b64encode instead of the normal b64encode

"""
Encode bytes using the URL- and filesystem-safe Base64 alphabet.

Argument s is a bytes-like object to encode.  The result is returned as a
bytes object.  The alphabet uses '-' instead of '+' and '_' instead of
'/'.
"""

closes https://github.com/back-to/liveproxy/issues/8